Move cloud_data_context_in_cloud_mode_with_datasource_pandas_engine
