+++ tests/data_context/conftest.py
@@ -8,8 +8,10 @@ from unittest.mock import Mock, PropertyMock, patch
 
 import pytest
 import requests
+from great_expectations.core.yaml_handler import YAMLHandler
 
 import great_expectations as ge
+from great_expectations import DataContext
 from great_expectations.data_context.store import GeCloudStoreBackend
 from great_expectations.data_context.store.ge_cloud_store_backend import AnyPayload
 from great_expectations.data_context.types.base import (
@@ -20,6 +22,8 @@ from great_expectations.data_context.util import file_relative_path
 from tests.integration.usage_statistics.test_integration_usage_statistics import (
     USAGE_STATISTICS_QA_URL,
 )
+
+yaml = YAMLHandler()
 
 
 @pytest.fixture()
@@ -794,3 +798,35 @@ def mocked_datasource_post_response(
         )
 
     return _mocked_post_response
+
+@pytest.fixture
+def cloud_data_context_in_cloud_mode_with_datasource_pandas_engine(
+    empty_data_context_in_cloud_mode: DataContext,
+    db_file,
+    mocked_datasource_get_response,
+):
+    context: DataContext = empty_data_context_in_cloud_mode
+    config = yaml.load(
+        f"""
+    class_name: Datasource
+    execution_engine:
+        class_name: PandasExecutionEngine
+    data_connectors:
+        default_runtime_data_connector_name:
+            class_name: RuntimeDataConnector
+            batch_identifiers:
+                - default_identifier_name
+        """,
+    )
+    with patch(
+        "great_expectations.data_context.store.ge_cloud_store_backend.GeCloudStoreBackend.list_keys"
+    ), patch(
+        "great_expectations.data_context.store.ge_cloud_store_backend.GeCloudStoreBackend._set"
+    ), patch(
+        "requests.get", autospec=True, side_effect=mocked_datasource_get_response
+    ):
+        context.add_datasource(
+            "my_datasource",
+            **config,
+        )
+    return context
